+++ tests/js/server/aql/aql-geo.js
@@ -38,7 +38,6 @@ const getQueryResults = helper.getQueryResults;
 const assertQueryError = helper.assertQueryError;
 const deriveTestSuite = require('@arangodb/test-helper').deriveTestSuite;
 const isEnterprise = require("internal").isEnterprise();
-let analyzers = require("@arangodb/analyzers");
 
 function makePolyInside(lon, lat) {
   // lon and lat are longitude and latitude ranges
@@ -208,6 +207,7 @@ function geoSuite(isSearchAlias, analyzerType) {
       withIndex = db._create(collWithIndex);
       withIndex.ensureIndex({type: "geo", geoJson: true, fields: ["geo"]});
       withView = db._create(collWithView);
+      let analyzers = require("@arangodb/analyzers");
       analyzers.save("geo_json", analyzerType, {}, ["frequency", "norm", "position"]);
 
       withIndex.save({ name_1: "name", "value": [{ "nested_1": [{ "nested_2": "foo123"}]}]});
@@ -281,7 +281,10 @@ function geoSuite(isSearchAlias, analyzerType) {
       db._drop(collWithoutIndex);
       db._drop(collWithIndex);
       db._drop(collWithView);
-      analyzers.remove("geo_json");
+/*
+      let analyzers = require("@arangodb/analyzers");
+      analyzers.remove(analyzerType);
+*/
     },
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -1105,344 +1108,6 @@ function geoSuite(isSearchAlias, analyzerType) {
   };
 }
 
-function BTS_470() {
-  return {
-    setUpAll: function () {
-      db._create("geo");
-      analyzers.save("geo_json", "geojson", {}, []);
-      db._createView("geo_view", "arangosearch", {
-       links: {
-          geo: {
-            fields: {
-              location: { analyzers: ["geo_json"] }
-            }
-          }
-        }
-      });
-  
-      let lat = 6.537;
-      let long = 50.332;
-      for (let x = 0; x < 500; ++x) {
-        let points = [];
-        for (let y = 0; y < 500; ++y) {
-          points.push({ location: { type: "Point", coordinates: [lat + x/1000, long + y/1000] } });
-        }
-        db.geo.save(points);
-      }
-      
-      db.geo.save({ location: { "type": "Polygon", "coordinates": [
-        [[ 37.614323, 55.705898 ],
-          [ 37.615825, 55.705898 ],
-          [ 37.615825, 55.70652  ],
-          [ 37.614323, 55.70652  ],
-          [ 37.614323, 55.705898 ]]
-      ]}});
-      db.geo.save({ location: {"type": "LineString", "coordinates": [
-        [ 6.537, 50.332 ], [ 6.537, 50.376 ]]
-      }});
-      db.geo.save({ location: { "type": "MultiLineString", "coordinates": [
-        [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-        [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-      ]}});
-      db.geo.save({ location: { "type": "MultiPoint", "coordinates": [
-        [ 6.537, 50.332 ], [ 6.537, 50.376 ],
-        [ 6.621, 50.332 ], [ 6.621, 50.376 ]
-      ]}});
-      db.geo.save({ location: { "type": "MultiPolygon", "coordinates": [
-        [[[ 37.614323, 55.705898 ],
-          [ 37.615825, 55.705898 ],
-          [ 37.615825, 55.70652  ],
-          [ 37.614323, 55.70652  ],
-          [ 37.614323, 55.705898 ]]],
-        [[[ 37.614, 55.7050 ],
-          [ 37.615, 55.7050 ],
-          [ 37.615, 55.7058 ],
-          [ 37.614, 55.7058 ],
-          [ 37.614, 55.7050 ]]]
-      ]}});
-
-      // sync view
-      db._query(`
-        LET lines = GEO_MULTILINESTRING([
-          [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-          [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-        ])
-        FOR doc IN geo_view SEARCH
-          ANALYZER(GEO_DISTANCE(doc.location, lines) < 100, "geo_json") OPTIONS {'waitForSync': true} return doc`);
-    },
-
-    tearDownAll: function () {
-      db._dropView("geo_view");
-      db._drop("geo");
-      analyzers.remove("geo_json");
-    },
-
-    test: function() {
-      const queries = [
-        [`
-          LET lines = GEO_MULTILINESTRING([
-            [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-            [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_DISTANCE(doc.location, lines) < 100, "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 5
-        ],
-        [`
-          LET lines = GEO_MULTILINESTRING([
-            [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-            [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_DISTANCE(doc.location, lines) < 3000, "geo_json")
-            FILTER doc.location.type != "Point"
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 3
-        ],
-        [
-          `
-          LET lines = GEO_MULTILINESTRING([
-            [[ 37.614323, 55.705898 ], [ 37.615825, 55.705898 ]],
-            [[ 37.614323, 55.70652 ], [ 37.615825, 55.70652 ]]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_DISTANCE(doc.location, lines) < 100, "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines)})
-          `, 2
-        ],
-        [
-          `
-          LET lines = GEO_MULTIPOINT([
-            [ 6.537, 50.332 ], [ 6.537, 50.376 ],
-            [ 6.621, 50.332 ], [ 6.621, 50.376 ]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_DISTANCE(doc.location, lines) < 100, "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 5
-        ],
-        [
-          `
-          LET lines = GEO_MULTIPOINT([
-            [ 6.537, 50.332 ], [ 6.537, 50.376 ],
-            [ 6.621, 50.332 ], [ 6.621, 50.376 ]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_DISTANCE(doc.location, lines ) < 3000, "geo_json")
-            FILTER doc.location.type != "Point"
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 3
-        ],
-        [
-          `
-          LET lines = GEO_MULTIPOINT([
-            [ 37.614323, 55.705898 ], [ 37.615825, 55.705898 ],
-            [ 37.614323, 55.70652 ], [ 37.615825, 55.70652 ]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_DISTANCE(doc.location, lines) < 100, "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines)})  
-          `, 2
-        ]
-      ];
-
-      queries.forEach( function (query_tuple, index) {
-        [query, expected] = query_tuple;
-        let actual;
-        try {
-          actual = db._query(query).toArray().length;
-          assertEqual(actual, expected);
-        } catch (err) {
-          print(`Actual: ${actual}, Expected: ${expected}, Index: ${index}`);
-          print(query);
-          fail(err);
-        }
-      }); 
-    }
-  }
-}
-
-function BTS_471() {
-  return {
-    setUpAll: function () {
-      db._create("geo");
-      analyzers.save("geo_json", "geojson", {}, []);
-      db._createView("geo_view", "arangosearch", {
-       links: {
-          geo: {
-            fields: {
-              location: { analyzers: ["geo_json"] }
-            }
-          }
-        }
-      });
-
-      let lat = 6.537;
-      let long = 50.332;
-      for (let x = 0; x < 500; ++x) {
-        let points = [];
-        for (let y = 0; y < 500; ++y) {
-          points.push({ location: { type: "Point", coordinates: [lat + x/1000, long + y/1000] } });
-        }
-        db.geo.save(points);
-      }
-
-      db.geo.save({ location: { "type": "Polygon", "coordinates": [
-        [[ 37.614323, 55.705898 ],
-          [ 37.615825, 55.705898 ],
-          [ 37.615825, 55.70652  ],
-          [ 37.614323, 55.70652  ],
-          [ 37.614323, 55.705898 ]]
-      ]}});
-      db.geo.save({ location: {"type": "LineString", "coordinates": [
-        [ 6.537, 50.332 ], [ 6.537, 50.376 ]]
-      }});
-      db.geo.save({ location: { "type": "MultiLineString", "coordinates": [
-        [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-        [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-      ]}});
-      db.geo.save({ location: { "type": "MultiPoint", "coordinates": [
-        [ 6.537, 50.332 ], [ 6.537, 50.376 ],
-        [ 6.621, 50.332 ], [ 6.621, 50.376 ]
-      ]}});
-      db.geo.save({ location: { "type": "MultiPolygon", "coordinates": [
-        [[[ 37.614323, 55.705898 ],
-          [ 37.615825, 55.705898 ],
-          [ 37.615825, 55.70652  ],
-          [ 37.614323, 55.70652  ],
-          [ 37.614323, 55.705898 ]]],
-        [[[ 37.614, 55.7050 ],
-          [ 37.615, 55.7050 ],
-          [ 37.615, 55.7058 ],
-          [ 37.614, 55.7058 ],
-          [ 37.614, 55.7050 ]]]
-      ]}});
-
-      // sync view
-      db._query(`
-      LET lines = GEO_MULTILINESTRING([
-        [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-        [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-      ])
-      FOR doc IN geo_view
-        SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 100), "geo_json") OPTIONS {'waitForSync': true}
-        RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-      `)
-    },
-    tearDownAll: function () {
-      db._dropView("geo_view");
-      db._drop("geo");
-      analyzers.remove("geo_json");
-    },
-    test: function() {
-      const queries = [
-        [
-          `
-          LET lines = GEO_MULTILINESTRING([
-            [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-            [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 100), "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 5
-        ],
-        [
-          `
-          LET lines = GEO_MULTILINESTRING([
-            [[ 6.537, 50.332 ], [ 6.537, 50.376 ]],
-            [[ 6.621, 50.332 ], [ 6.621, 50.376 ]]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 3000), "geo_json")
-            FILTER doc.location.type != "Point"
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 3
-        ],
-        [
-          `
-          LET lines = GEO_MULTILINESTRING([
-            [[ 37.614323, 55.705898 ], [ 37.615825, 55.705898 ]],
-            [[ 37.614323, 55.70652 ], [ 37.615825, 55.70652 ]]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 100), "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines)})
-          `, 2
-        ],
-        [
-          `
-          LET lines = GEO_MULTIPOINT([
-            [ 6.537, 50.332 ], [ 6.537, 50.376 ],
-            [ 6.621, 50.332 ], [ 6.621, 50.376 ]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 100), "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 5
-        ], 
-        [
-          `
-          LET lines = GEO_MULTIPOINT([
-            [ 6.537, 50.332 ], [ 6.537, 50.376 ],
-            [ 6.621, 50.332 ], [ 6.621, 50.376 ]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 3000), "geo_json")
-            FILTER doc.location.type != "Point"
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines )})
-          `, 3
-        ], 
-        [
-          `
-          LET lines = GEO_MULTIPOINT([
-            [ 37.614323, 55.705898 ], [ 37.615825, 55.705898 ],
-            [ 37.614323, 55.70652 ], [ 37.615825, 55.70652 ]
-          ])
-          FOR doc IN geo_view
-            SEARCH ANALYZER(GEO_IN_RANGE(doc.location, lines, 0, 100), "geo_json")
-            RETURN MERGE(doc, { distance: GEO_DISTANCE(doc.location, lines)})
-          `, 2
-        ]
-      ];
-      
-      queries.forEach( function (query_tuple, index) {
-        [query, expected] = query_tuple;
-        let actual;
-        try {
-          actual = db._query(query).toArray().length;
-          assertEqual(actual, expected);
-        } catch (err) {
-          print(`Actual: ${actual}, Expected: ${expected}, Index: ${index}`);
-          print(query);
-          fail(err);
-        }
-      }); 
-    }
-  }
-}
-
-function BTS_471Suite() {
-  let suite = {};
-  deriveTestSuite(
-    BTS_471(),
-    suite,
-    "BTS_471"
-  );
-  return suite;
-}
-
-function BTS_470Suite() {
-  let suite = {};
-  deriveTestSuite(
-    BTS_470(),
-    suite,
-    "BTS_470"
-  );
-  return suite;
-}
-
 function arangoSearchVPackGeoSuite() {
   let suite = {};
   deriveTestSuite(
@@ -1485,8 +1150,6 @@ function searchAliasS2GeoSuite() {
 
 jsunity.run(arangoSearchVPackGeoSuite);
 jsunity.run(searchAliasVPackGeoSuite);
-jsunity.run(BTS_470Suite)
-jsunity.run(BTS_471Suite)
 /*
 jsunity.run(arangoSearchS2GeoSuite);
 jsunity.run(searchAliasS2GeoSuite);
